backend/requirements.txt
asgiref==3.11.0
async-timeout==5.0.1
Django==5.2.9
django-cors-headers==4.9.0
django-debug-toolbar==6.1.0
django-environ==0.12.0
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
hiredis==3.3.0
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
redis==7.1.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
